feat: Add `error handling` and `request validation` with `zod` in all routes

+++ src/app/api/v1/notes/[noteId]/route.ts
@@ -1,5 +1,8 @@
+import errorHandler from "@/app/errors/errorHandler";
+import { NotFoundError } from "@/app/errors/errors";
 import { PrismaClient } from "@prisma/client";
 import { NextResponse } from "next/server";
+import { z } from "zod";
 
 const prisma = new PrismaClient();
 
@@ -9,54 +12,73 @@ interface IParams {
   };
 }
 
-// Route to edit a Note
-export async function PUT(req: Request, { params }: IParams) {
-  const { noteId } = await params;
-  const { title } = await req.json();
+const noteIdSchema = z.object({
+  noteId: z
+    .string()
+    .regex(/^[0-9]+$/gm, "Only numbers are accepted")
+    .transform((arg) => Number(arg)),
+});
 
-  const note = await prisma.note.update({
-    where: { id: Number(noteId) },
-    data: { title: title },
+// Route to edit / rename a Note
+export async function PUT(req: Request, { params }: IParams) {
+  const titleSchema = z.object({
+    title: z.string().nonempty(),
   });
 
-  if (!note) {
+  try {
+    const { noteId } = noteIdSchema.parse(await params);
+    const { title } = titleSchema.parse(await req.json());
+
+    // checks if the note exists
+    const checkNote = await prisma.note.findFirst({
+      where: { id: noteId },
+    });
+
+    if (!checkNote) {
+      throw new NotFoundError({ message: "Couldn't find especified note" });
+    }
+
+    const note = await prisma.note.update({
+      where: { id: noteId },
+      data: { title: title },
+    });
+
     return NextResponse.json(
-      { message: "Note not found" },
+      { updatedNote: note },
       {
-        status: 404,
+        status: 200,
       },
     );
+  } catch (error) {
+    return errorHandler(error);
   }
-
-  return NextResponse.json(
-    { updatedNote: note },
-    {
-      status: 200,
-    },
-  );
 }
 
 // Route to delete a Note
 export async function DELETE(_req: Request, { params }: IParams) {
-  const { noteId } = await params;
+  try {
+    const { noteId } = noteIdSchema.parse(await params);
 
-  const note = await prisma.note.delete({
-    where: { id: Number(noteId) },
-  });
+    // checks if the note exist
+    const checkNote = await prisma.note.findFirst({
+      where: { id: noteId },
+    });
 
-  if (!note) {
+    if (!checkNote) {
+      throw new NotFoundError({ message: "Couldn't find especified note" });
+    }
+
+    const note = await prisma.note.delete({
+      where: { id: Number(noteId) },
+    });
+
     return NextResponse.json(
-      { message: "Note not found" },
+      { deletedNote: note },
       {
-        status: 404,
+        status: 200,
       },
     );
+  } catch (error) {
+    return errorHandler(error);
   }
-
-  return NextResponse.json(
-    { deletedNote: note },
-    {
-      status: 200,
-    },
-  );
 }
